Added default material for quickly identifying different objects with the same mesh (temporary). Created a default object for testing quest completion functionality which uses said mesh.
Updated the UI to have a task list of up to 5 tasks - These are removed when the task is completed and the list is repopulated. Updated the task struct to have all ID's be of type Name to prevent type conversions. Fixed an issue where tasks could be in the available and completed lists at the same time due to not checking the completed list properly on task completion. Created a generic task load function for the UI and added specific implementations for each task bullet. Fixed an issue with out of index errors due to not enough tasks being available. Fixed an issue with the task UI not being able to access the task manager as when created the reference within the UI wasn't given the manager to reference.

Gave the player a take damage function. Created a game over level which the player is taken to when they have 0 hp or less. Fixed an issue where the health bar refilled due to changing how damage is taken. Fixed an issue where the player took no damage from changing states.

Gave the player the ability to interact with objects. Changed the size of the hitbox to be more appropriate for the game style.